requirements.txt
# Core dependencies for devin.cursorrules tools
python-dotenv>=1.0.0
playwright>=1.40.0
html5lib>=1.1

# LLM API clients
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.7.0

# Web scraping and search
duckduckgo-search>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Image processing (for screenshot tools)
Pillow>=10.0.0

# Utilities
tqdm>=4.66.0
